Extract common code out of ArrayDisplayColumn and MapArrayDisplayColumn into AbstractExcelDisplayColumn
